requirements.txt
agno==1.5.4
annotated-types==0.7.0
anyio==4.9.0
baidusearch==1.0.3
bcrypt==4.0.1
beautifulsoup4==4.13.4
bidict==0.23.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-genai==1.13.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.10.0
lancedb==0.22.0
limits==5.1.0
lxml==5.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.5
ollama==0.4.8
openai==1.82.0
overrides==7.7.0
packaging==25.0
passlib==1.7.4
pgvector==0.4.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg2-binary==2.9.10
pyarrow==19.0.1
python-jose[cryptography]
# pyasn1
# pyasn1_modules==0.4.2
pycountry==24.6.1
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
# python-jose==3.4.0
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
# rsa==4.9.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
slowapi==0.1.9
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
tomli==2.2.1
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
